backend/requirements.txt
# Server
pinecone-client
langchain-pinecone
langchain-google-genai

# Core RAG & Agent framework
langchain
langchain-community
langchain-openai
langchain-huggingface

# Vector Database
chromadb

# Text/Code Splitting
langchain-text-splitters

# Embedding Model
sentence-transformers

# Git Loader
GitPython

# Environment variables
python-dotenv

einops
